% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{calc_age_matrix}
\alias{calc_age_matrix}
\title{Calculate age weight matrix for AST}
\usage{
calc_age_matrix(n_age, par_age = 1)
}
\arguments{
\item{n_age}{Number of age groups.}

\item{par_age}{Age smoothing parameter (omega, default 1).}
}
\value{
Square age weight matrix.
}
\description{
Exponential decay: \verb{W_A(i,j) = 1 / exp(omega * |i-j|)}
}
\keyword{internal}
