% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{calc_time_matrix}
\alias{calc_time_matrix}
\title{Calculate temporal weight matrix for AST}
\usage{
calc_time_matrix(min_year, max_year, par_time = 2)
}
\arguments{
\item{min_year}{Start year.}

\item{max_year}{End year.}

\item{par_time}{Temporal smoothing parameter (lambda, default 2).}
}
\value{
Square temporal weight matrix.
}
\description{
LOESS-style cubic power weighting: \verb{W_T(i,j) = (1 - (|i-j|/T)^lambda)^3}
}
\keyword{internal}
