% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_suggests}
\alias{check_suggests}
\title{Check that a package is available}
\usage{
check_suggests(pkg, reason = NULL)
}
\arguments{
\item{pkg}{Package name.}

\item{reason}{Why it is needed.}
}
\value{
\code{TRUE} invisibly, or aborts.
}
\description{
Check that a package is available
}
\keyword{internal}
