% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{compute_bayes_residuals}
\alias{compute_bayes_residuals}
\title{Compute residuals from Bayesian fit (posterior means)}
\usage{
compute_bayes_residuals(fit_bayes)
}
\description{
Compute residuals from Bayesian fit (posterior means)
}
\keyword{internal}
