% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{compute_bootstrap_ui}
\alias{compute_bootstrap_ui}
\title{Compute bootstrap uncertainty intervals}
\usage{
compute_bootstrap_ui(fit, adjacency, config, dentition, n_boot, dmft_max)
}
\description{
Compute bootstrap uncertainty intervals
}
\keyword{internal}
