% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{detect_region_col}
\alias{detect_region_col}
\title{Detect the region name column in an sf object}
\usage{
detect_region_col(sf_obj, config)
}
\description{
Detect the region name column in an sf object
}
\keyword{internal}
