% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{dmft_load_shapefile}
\alias{dmft_load_shapefile}
\title{Load a shapefile for spatial analysis}
\usage{
dmft_load_shapefile(shapefile_path)
}
\arguments{
\item{shapefile_path}{Path to \code{.shp} file (or any format \code{\link[sf:st_read]{sf::st_read()}}
supports).}
}
\value{
An \code{sf} object.
}
\description{
Load a shapefile for spatial analysis
}
