% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{extract_year_trend}
\alias{extract_year_trend}
\title{Extract temporal trend from year random effects}
\usage{
extract_year_trend(fit, config)
}
\description{
Extract temporal trend from year random effects
}
\keyword{internal}
