% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{predict_bayes_grid}
\alias{predict_bayes_grid}
\title{Predict from Bayesian model for a full grid (posterior means)}
\usage{
predict_bayes_grid(fit_bayes, grid, config)
}
\description{
Predict from Bayesian model for a full grid (posterior means)
}
\keyword{internal}
