% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{predict_lmer_grid}
\alias{predict_lmer_grid}
\title{Predict from lme4 model for a full grid}
\usage{
predict_lmer_grid(fit, grid, config, dentition)
}
\description{
Predict from lme4 model for a full grid
}
\keyword{internal}
