% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{prepare_lmer_data}
\alias{prepare_lmer_data}
\title{Prepare data for lme4 mixed model}
\usage{
prepare_lmer_data(data, adjacency, dentition, config)
}
\description{
Prepare data for lme4 mixed model
}
\keyword{internal}
