% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_bayes.R
\name{prepare_stan_data}
\alias{prepare_stan_data}
\title{Prepare Stan data list from model data}
\usage{
prepare_stan_data(mdata, config)
}
\description{
Prepare Stan data list from model data
}
\keyword{internal}
