% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_linear}
\alias{project_linear}
\title{Project using linear extrapolation with damping}
\usage{
project_linear(trend, n_forecast, adjustment, damping_factor, dmft_max, seed)
}
\description{
Project using linear extrapolation with damping
}
\keyword{internal}
