% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_dmft}
\alias{theme_dmft}
\title{Default ggplot2 theme for DMFT plots}
\usage{
theme_dmft()
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
Default ggplot2 theme for DMFT plots
}
