% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmft-package.R
\docType{package}
\name{dmft-package}
\alias{dmft}
\alias{dmft-package}
\title{dmft: Bayesian Hierarchical Modeling of Dental Caries (DMFT/dmft)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements GBD-style Bayesian hierarchical models for estimating and projecting dental caries burden (DMFT/dmft indices) at subnational level. Combines BYM2 spatial random effects, RW2 temporal trends, and RW1 age effects with Gaussian meta-regression. Supports any country given study-level data and a regional shapefile. Provides a complete pipeline from data cleaning through model fitting, diagnostics, projections, and publication-ready visualizations. Two inference backends are supported: INLA (fast approximate Bayesian) and Stan (full MCMC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/choxos/dmft}
  \item Report bugs at \url{https://github.com/choxos/dmft/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Sofi-Mahmudi \email{ahmad.sofimahmudi@gmail.com} (\href{https://orcid.org/0000-0002-2460-2394}{ORCID}) [copyright holder]

}
\keyword{internal}
